Removes check fo malloc.h on FreeBSD
The inclusion of malloc.h results in a failed build (if __STD__ is
defined, otherwise stdlib.h is included). malloc and friends are
defined in stdlib.h, per C standard.

An undefined HAVE_MALLOC_H is sufficient for clean builds on FreeBSD,
but maybe all references to malloc.h should be removed?