MODIFIED: Memory management for temporary strings used internally and through the foreign language interface functions such as PL_get_nchars(). Such strings used to be managed (by default) in a _ring_ of 16 strings, reusing old buffers. Recent versions used memory mapping for long strings and unmapped these. This exposed a problem that input strings could be incorrectly reused early.
Together with Keri Harris we designed and implemented a new stack based
management for such strings. Temporary strings are now reclaimed on exit
(or failure) of a foreign predicate that created them. In addition,
there is this C block ({...}) that may be used to scope strings:

    PL_STRINGS_MARK();
    <code>
    PL_STRINGS_RELEASE();

A new flag `string_stack_tripwire` can be set to get a disagnostic message
printed if the string stack reaches some depth limit.  This may hint at
using the above macros to reduce resource usage.

In addition a number of PL_* functions that used to step the string ring
now use the above macros to reduce resource usage.