DOC: Improved intro text for library(yall) intro
I have added more explanations regarding the concept of "closure" as used in Prolog and extended the description of the rest. It feels better and is hopefully clearer to a newcomer, but in the end it's matter of taste. I was unable to justify the text to 73 columns though, I just kept to this as close as possible,