DOC: Rewrote "Overview of the Debugger"
I extended the text/command line debugger overview to add more detail and examples for newer users.  It is pretty much a rewrite, but I reused some of the text from the original.

The feedback threads on discourse for it are here: https://swi-prolog.discourse.group/t/learning-how-to-use-the-prolog-command-line-debugger-try-this-new-overview/3225 and here: https://swi-prolog.discourse.group/t/feedback-please-on-proposed-update-to-tutorials-beginner-debugger/3216.

I wrote it as a markdown file in attempt to avoid learning latex :-).  Seemed to work pretty well (writeup here: https://swi-prolog.discourse.group/t/using-markdown-with-pldoc-to-generate-manual-content/3257/2), but required using the latext \input command to get the input into the master overview.doc and and update to CMakeLists.txt to build.

I confirmed that the HTML builds and looks right.