FIXED: Possible race condition in retract/1 generation management.
This patch ensures that when a clause is considered visible in some thread
based on `GD->_generation`, this remains to be the case.  Before, the global
generation may be updated before the clause generation is updated, which may
trigger a thread to consider the clause visible while it is no longer.

Finally diagnosed by Keri Harris.