ADDED: Initial SICStus 4 emulation (expects_dialect(sicstus4)).
This is separate from the existing SICStus emulation
(expects_dialect(sicstus)), which emulates almost only SICStus 3
features. The few SICStus 4-only features that were emulated by the
sicstus dialect have been moved to the sicstus4 dialect.

SICStus 3 and 4 are incompatible enough in some aspects that they cannot
be emulated properly as a single dialect. See this discussion:
https://swi-prolog.discourse.group/t/expects-dialect-sicstus-how-to-handle-sicstus-3-vs-4-api-conflicts/3381

The SICStus 4 emulation currently supports all of library(system) and
library(timeout), parts of library(lists) and library(terms), and a few
builtins. library(sockets) is not emulated yet, because its API has
changed completely between SICStus 3 and 4. library(arrays) has been
removed entirely in SICStus 4, so it is not emulated by the sicstus4
dialect.