Add custom goal representation to SICStus block emulation
This makes the toplevel, frozen/2, etc. render blocked goals the same
way as SICStus, i. e. as simply the original predicate call. This is
especially useful for programmatically checking if a certain blocked
goal exists on a variable.

This requires reimplementing most of the logic manually using attvars,
so that attribute_goals//1 can be implemented. With the previous
emulation based on freeze/when, blocked goals would be rendered as the
generated freeze/when calls, with no way to customize this.