DEBUG: Early/better internal debug output
Initialize the global debug_level as soon as possible if "-d" is passed
first on the command line, to avoid missing early debug lines. Also,
track use of the DEBUG() macro on a per-thread basis and use Serror
as the default output stream when inside a DEBUG clause, so that
statements that mix e.g. Sdprintf() and write_term() will get printed
to a single output. Finally, prefix debug lines with the debug channel
and the file/line source.