Converting shared code labels to helper segments
This changes all cross-VMI jumps into jumps to helper
pseudoinstructions, or VMHs, defined in much the same way VMIs
are, though with C-level callsite arguments. Note, however, that
at present, the arguments are nonfunctional; values are still
being passed via SHAREDVARS. Also, moved the ancillary chunks from
the end of PL_next_solution() into pl-vmi.c as VMH helpers, along
with some of the support macros.

Now, the only ways to exit a VMH/VMI are via macros, including the
new VMH_GOTO() which takes a VMH name and an argument list, and
SOLUTION_RETURN() which is (currently) just a wrapper for return.