Register-mode functions
This is a still-fairly-experimental modification to the VMI functions
introduced in the previous commit. It uses a GCC extension,

register int foo asm("rax");

to specify that a given file-level variable should always be considered
to be stored in the given register, and that the compiler should never
use that register for any other purpose. REGISTERS and LD are passed
that way in PL_next_solution(), effectively creating a custom, one-off
calling convention for this file only.