Debug wildcard and complement; DEBUG_IF
The -d command line option now takes an optional asterisk (*) on the
end of a debug flag name to enable all flags that begin with that name.
Also, flag names can be prefixed with a caret (^) to disable that flag;
only particularly useful in combination with the wildcard, to silence
chatty channels. For example,

    swipl -d MSG_SHIFT*,^MSG_SHIFT_FRAME

will enable all MSG_SHIFT channels except for MSG_SHIFT_FRAME.

On the internal side, we now have the DEBUG_IF(cond, channel) macro,
which functions much like DEBUG(channel, statements) except it provides
more control over the conditions, and it avoids statements in a macro.
As a result, the statement

    DEBUG(MSG_CHANNEL, Sdprintf(...));

is (for a named channel, at least) equivalent to

    int debug_active = DEBUGGING(MSG_CHANNEL);
    DEBUG_IF(debug_active, MSG_CHANNEL)
    { Sdprintf(...); }

which itself is roughly equivalent to:

    #if O_DEBUG
    int debug_active = DEBUGGING(MSG_CHANNEL);
    if (debug_active)
    { ENTER_DEBUG(MSG_CHANNEL);
      Sdprintf(...);
      EXIT_DEBUG(MSG_CHANNEL);
    }
    #endif

This commit also includes some improvements to the GC debug messaging.