BUILD: Added self-contained PGO process to CMake.
This change makes it possible to run a PGO build with a single
invocation of the build process. To do so, use the "PGO" configuration
in CMake rather than "Debug" or "Release", which will:

1) Compile all libswipl objects with profile generation enabled. This
   uses auxiliary target names with "-pgo-instrumented" appended.
2) Link the executable "swipl-pgo-instrumented" in src, using these.
3) Use swipl-pgo-instrumented to create boot.prc (since no other swipl
   exists yet). Clear the PGO-data directory, then run the benchmark.
4) Build the libswipl objects again, using the PGO data generated.
5) Link the final swipl executable from these.

The dependencies are set up such that if any source file changes,
ninja will recompile the instrumented version of the affected file(s)
and then restart the process from step 3, recompiling ALL libswipl
objects using the new PGO data.

Tested with gcc 9/10 and clang 10/11.