FIXED: missing GMP exports, -Wmissing-prototypes
This addresses issues exposed by adding -Wmissing-prototypes to the
build, hacked in as `cmake -DCHECK_PROTOTYPES=1`. This flag emits a
warning when any non-static function is defined without first declaring
a separate prototype, which usually means that either a header hasn't
been included, or that a function that should have been static wasn't
marked as such. The Linux compile is clean now, after having:

- added `static` to functions not called outside the file
- completed incomplete header prototypes, usually by adding (void)
- removed old functions no longer called anywhere
- included gmp.h before SWI-Prolog.h, to enable the missing GMP exports

The Windows build still has some missing-prototype warnings because of
various #ifdefs, but I'm not addressing those right now.