MODIFIED: LDFUNC refactor, part 1 (automated)
This is the first of two commits that applies a refactor across the
entire SWI-Prolog codebase, changing the syntax of passing LD from
one function to another in internal library code. More details on
the refactor are in the following commit, but since this touches
so much code and to avoid having to manually verify a 20,000+ line
diff, the bulk of the changes in this commit are automated by a Perl
script, saved here (for this commit only) as src/ld-refactor.perl,
with the output of the run saved in src/ld-refactor.out. This way,
the script itself can be audited for correctness, and only the
special cases in the following commit need to be checked manually.