Reimplement SICStus block emulation using attvars directly
The main advantage over the previous frozen/when-based implementation
is that blocked goals now have a nice goal representation in frozen,
copy_term, etc. - a blocked goal is now simply displayed as the
original call, not as a frozen or when call. This also improves
compatibility with code written for SICStus that looks for specific
blocked goals in the output of frozen/copy_term.

The internal attribute format is intentionally very similar to that of
when/2. This also allows reusing the native '$suspend'/3 predicate for
updating a collection of delayed goals attached to a variable.