MODIFIED: LDFUNC refactor, part 2 (manual)
This is the second of two commits that applies a refactor across the
entire SWI-Prolog codebase, changing the syntax of passing LD from
one function to another in internal library code. The bulk of the
changes in this refactor were automated via Perl script in the previous
commit; see that commit message for more details on that.

Previously, functions that wanted to access LD (the local data
structure) could either use the GLOBAL_LD macro, declare the
GET_LD/PRED_LD macros early in the function, or declare ARG_LD/ARG1_LD
somewhere in the function parameter list. Calling these functions
required either declaring a macro (usually in pl-ldpass.h) that would
add the LD argument to any call to the function, or adding a
PASS_LD/PASS_LD1 annotation to each callsite.

The new style revolves around a new macro called LDFUNC, defined in
src/pl-builtin.h; see that file for details on its use. PASS_LD and
its variants are no longer required at callsites; all LDFUNC-augmented
functions have macro definitions that add the necessary arguments
to the function call. Additionally, name-mangling is performed on
function names, so it is no longer required to add `__LD` to a function
name to keep it from colliding with externally-visible API functions.

This commit also adds the API_STUB macro, which ensures that API
function definitions do *not* have their names mangled, along with
including a GET_LD declaration automatically.

For more information, see the following post:

https://swi-prolog.discourse.group/t/changes-to-ld-handling-in-library-code/4067