Avoid rerunning defatom and mkvmi for no reason
Previously, defatom/mkvmi avoided updating the output file (and its
mtime) if the newly generated data is identical to the existing
contents of the file. This caused problems if the input files have
changed, but in a way that doesn't change the output data. In that
case, defatom/mkvmi is called on every build - the output is considered
outdated as its mtime is older than that of the inputs, but running
defatom/mkvmi doesn't update the mtime, so on the next build it's still
considered outdated. This would repeat until the inputs change in a way
that actually changes the output.

To fix this, defatom/mkvmi now unconditionally updates the output file.
This fixes the problem above, at the cost of sometimes causing an
unnecessary rebuild of things that depend on the outputs (but this will
only happen once per change, not on every build as with the previous
problem).