FIXED: Issue#105: dif/2 wrong result.
This patch is a major rewrite for dif/2.  Where the previous version contained
_or nodes_ that were considered satisfied based on a count, the new version
verifies the individual unifications of the _unifier_ produced by unifiable/3.
The count fails, both too low and too high, as a result of multiple unifications
in a single unification as well as subterm sharing in the input terms.