MODIFIED: write_term/3 option `brace_terms` behavior was the inverse of the documentation.  As the docs describe what one would expect and is consistent with YAP, the implementation is fixed.
We do not expect this option to be frequently used.  Code using `brace_terms(true)`
must be updated to use `brace_terms(false)`.   If you need to do a runtime check
to beor conditional compilation compatible with older versions, one can use this
condition to verify the correct new behavior.

    term_string({a}, S, [brace_terms(false)]), S == "{}(a)"