ADDED: on_signal/3 now takes value 'ignore'
This allows for requesting the SIG_IGN disposition, and also unifies the
SIGPIPE behavior with the other initially-prepared signals. Technically,
this could be considered a backwards-incompatible change. However, for
this to actually affect someone's working code, they would have to

1. Redefine the system predicate ignore/1,
2. Call `on_signal(Sig, _, ignore)`, AND
3. Expect `ignore(Signal)` to do something besides ignoring Signal;

and in that case I think we can all agree they've brought this trouble
upon themselves. :)

Also adds SIGPIPE's default disposition to the docs.