ENHANCDD: better support for platforms without signals.
This adds an alternate alerting path when running on a non-Windows
system without sig_alert, using a C11 mutex and condition variable.
Obviously this loses the "interrupt blocking system call" semantics,
but it allows thread_sigwait to park itself while waiting for another
thread's signal.

A few tests needed changing:
1. When running without signals on non-Windows, popen(cat-2) will set
   SIGPIPE to ignored manually, since there's no other way to avoid
   that signal.
2. The signal_nested tests needed an extra call port so they could
   proactively check for virtual signals.
3. The D_BREAK tests simply don't work. It's slightly too big of a
   can of worms for me to look into for this commit, so I simply
   disabled that test suite on signal-free systems.

A couple package tests needed adjustments as well:

1. clib:test_af_unix was using an interrupting signal to stop the server
   thread, so I've changed it to auto-exit on signal-free systems.
2. ssl:ssl_certificates was closing the client before reading the
   server's response, which caused a SIGPIPE on the server, so I added
   a read-and-check to the test to drain the buffer.

For testing purposes, the no-signals build can be approximated on a
system built with signal support by passing the --no-signals and
--sigalert=0 arguments to swipl on the command line.