Do not install WIN32 dlls if WIN32_DLLS=""
This is a bit of a hack. I am currently preparing a PKGBUILD for MSYS2/MinGW, and the package installation tries to overwrite the existing zlib1.dll et al. of the MSYS2 system. So whenever WIN32DLLS is set to "" during the cmake build, install_dll does nothing. I think that on the long run, the job of install_dll should be moved to cpack.