PORT: Fixed rationalize/1 and cmpFloatNumbers() to fail due to gcc optimization using internal wide float registers rather than comparing 64 bit doubles.
Not entirely sure this is our mistake or within GCC's claim to work
with the extended float registers.   Bottom line is that comparing
a double (from memory, so 64 bits) to the result of a computation
may return non-equal even if the extended float when forced to 64
bits is equal.  This problem surfaced in i386 using gcc 12.2.0.