MODIFIED: Writing ``\`` quoted without `character_escapes` should not double the ``\``. This bug was introduced with cbe691f8af61ccaf466147868a316b96c110fedb to "fix" ``term_to_atom('a\\b', A)`` to produce ``A = 'a\\b'`` (which is quoted again by the toplevel to produce ``A = '\'a\\\\b\''``.
The issue is that term_to_atom/2 using PL_write_term() with
``PL_WRT_QUOTED``, but __not__ ``PL_WRT_CHARESCAPES``. There is now a
new flag ``PL_WRT_NO_CHARESCAPES``. If either flag is given, we
explicitly enable or disable character escapes. If none is given, the
behaviour depends on the global (module `user`) value of the
`character_escapes` flag.

This implies that e.g. ``term_to_atom('a\nb', A)`` used to produce a
quoted atom with a plain newline, while in the new version the newline
is escaped as ``\n``.

Many foreign libraries using PL_write_term() or PL_get_chars() using
CVT_WRITE did not use character escapes while they do now. As long as
the output is read by Prolog the result is the same.