MODIFIED:  Renamed exception processing hook user:prolog_exception_hook/4 into prolog:prolog_exception_hook/5, providing an additional argument that contains the current setting of the `debug` flag.  Note that we cannot ask for this flag as the hook runs in "nodebug" mode.   As the change is incompatible anyway, we also move the hook to the `prolog` module where the modern system hooks live.
The new argument is used by library(prolog_stack) to make catch/3 behave
as catch_with_backtrace/3 when in debug mode if the "guard" `debug` is
defined (default).   If this is not desired, use

```
:- use_module(library(prolog_stack)).
:- retractall(prolog_stack:guard(debug)).
```

Code that use the old hook must rename the predicate and add the `Debug`
argument.   Code that wishes to be compatible with old and new Prolog can
implement both hooks.