SECURITY: CVE-2023-45853
Upstream bug in the _minizip_ library used to create and load
Prolog saved states.   Not very relevant to SWI-Prolog as it
unlikely that SWI-Prolog is used in a way that provides
uncontrolled access to the involved parameters.

Upstream message:

Reject overflows of zip header fields in minizip.

This checks the lengths of the file name, extra field, and comment
that would be put in the zip headers, and rejects them if they are
too long. They are each limited to 65535 bytes in length by the zip
format. This also avoids possible buffer overflows if the provided
fields are too long.