INSTALL: Allow installing in public directory without symlink
For some installations we want the executables in another directory
than the "home".  Normally these were POSIX systems and we would symlink
from the executable to the real location in the self-contained binary.

But, also some Windows installations such as MSYS2 and Conda (and probably
Cygwin) want this.   Therefore we added SWIPL_INSTALL_AS_LINK which, if
OFF, compiles the relative path to the anticipated home into the binary
such that we can install the .exe in a separate directory (for Windows
along with the DLLs needed to run them).