CLEANUP: Move installation locations to cmake/Locations
Moving it into a separate file keeps the main file short and it needs to be before
`Ports` because we need to act on them.   Possibly though, we need part of
the locations before and part after the Ports?  As is, that does not seem
to be the case.   Alternatively we could move part of this into the Port
files?