MODIFIED: Install each extension in library/ext/dir
This allows adding and removing extensions without rebuilding the
library index and clarifies the source of files.  This was introduced
to simplify adding and deleting optional components in e.g., the
Windows installer or package managers that split SWI-Prolog into
multiple packages such as used by Debian and planned for Conda.