FIXED: Use snapshot/1 when computing exports
This is used to update the autoload library as well as several other
IDE related tasks.   For this we read and expand the file until we
find the first non-directive.   This may however leave traces in
dynamic predicates.   Using snapshot/1 avoids that.

It is a partial work around.  There could be other side-effects.  Should
probably be considered experimental.