FIXED: Qlf file generation for ``:- include(File)``
The include statements can come in more places, even in the middle of
a predicate.   This changes the identifier from 'I' to 'L' to make the
statement unique in all contexts and defines handling the include
statements.

Updates the qlf save vesion to 69.  Old valid version 68 files can
still be loaded.