FIXED: Thread I/O streams may be invalid
If a thread is created using thread_create/3 with the inherit_from(Thread)
option or using PL_thread_attach_engine() (which inherits from `main`), we
may copy `current_input` or `current_output` from a temporary redirect.

The reference counting when copying the streams was broken, possibly
causing memory corruption.  This is improved, but not 100% safe.  However,
we now set the `current_input` and `current_output` to `user_input` and
`user_output`, which typically never change.