FIXED: Handle source files with epoch time stamp
Flatpak seems to set the modification time of all times in the sandbox
to the epoch (0).  This was used as a special value to indicate the
source file was not a file.   We now keep a seperate flag and use the
_integer_ 0 to indicate something is not a source file in the Prolog
world.   It is probably better to use something else, but this avoids
type issues in possibly unknown dependent code.