FIXED: Do not wait longer than 30 seconds after a crash.
This deals with situations where error reporting and cleanup
hangs due to a deadlock.

Normally, after a fatal crash report and if the system is connected to
a terminal, it waits for user input such that the user can attach a
debugger to analyse the state.   We now wait at max 30 seconds to avoid
indefinite waiting if there is no human in the loop after all.