APPLE: Add code signing for creating the .dmg image
The image is signed if ``$CODESIGN_ID`` is set to the certificate id
to be used with ``codesign -s``.   Unfortunately, this is not the whole
story.  As a result, we can no longer load shared object that are not
part of the disk image, which means we cannot use Xquartz, Python or Java.