FIXED: build order for dependencies in packs
I think the order is in the wrong way. Say, pack a depends on pack b, and both have post-installation scripts. Then, pack b's script should be run before pack a's script. This does not seem to be the case here. If I reverse the order, it is running fine.

There's probably a better way to fix this (e.g., in build_order).