ADDED: write_term/2: option radix(Radix).
Experimental.  Under discussion as PIP-105.  The current implementation
writes, unlike XSB, the radix prefix.  We either need more options
or we should consider another alternative such as integer_format(+Format)

As this patch cleans up a lot of the code regarding numeric output and
most will be needed anyway, it is merged despite the fact that the
radix option may vanish eventually.