ADDED: write_term/2 options format_float(+Format) and format_integer(+Format)
Using a format seems a more sensible way to control writing numbers from
write_term/2.  This patch removes the previously introduced radix(+Radix)
option.  If we want this back we will implement it on top of
format_integer(+Format).

This patch also contains a lot of small code cleanup and some optimization
to low-level I/O.