ADDED: Prolog flag `unknown_option` with values {ignore,warning,error}.
As a result of PIP-105.   May change, but unlikely as their was clear
agreement.

This flag disables the `iso` flag from reporting errors on unknown options
because this breaks large parts of SWI-Prolog's libraries.   The user can
now explicitly enable this as well as make the system pring warnings to
investigate possible issues.