MODIFIED: pack_install/2 option autoload(true)
If present, the pack is installed as autoload pack regardless of
`autoload` option in the meta-data.  If the option is not present,
the system asks the user to install as autoload library if the
meta-data contains autoload(true).

If autoloading is not enabled, the ``INDEX.pl`` file that may appear
in the pack is removed.

This patch also adds ``swipl pack install --autoload ...``