ENHANCED: No longer trigger AGC from CGC
We used to trigger this if there are a fair amount of clauses that
have been erased for which we have a clause reference.   As we remove
these clauses anyway from the indexes, they only cost memory and thus
leaving the decision to normal AGC seems fine.

In due time we should probably consider atom sizes when considering
AGC.