ADDED: Yield based debugger support
This patch introduces a large number of changes to the VM as well
as the debugger code.  It allows interacting with the Prolog debugger
using _yield_, i.e., instead of calling a hook when a Prolog port is
passed it makes PL_next_solution() return.  The environment can examine
the state, make calls on this Prolog engine and finally set how the
session must be continued and resume PL_next_solution().

The interaction is similar to user:prolog_trace_interception/4, used
to drive the GUI debugger using callbacks.  Given this, it becomes
possible to write a debugger for the WASM version for using in
the browser.  This can either be a traditional console debugger or
a graphical debugger similar to the native debugger.

This patch reorganises notably the backtracking and exception handling
code in the VM, splitting these large pieces of code into many VM
helper functions.  It also reorganises several parts of the builtin
debugger to facilitate code reuse between the callback and yield
based debuggers.

This is a squash commit of the ``trace-yield-state`` branch.