FIXED: We must insert I_CONTEXT in multifile predicates
Multifile predicates are typically installed in modules using
`target:head :- body`, where body is executed in the containing
module.  This is achieved using an I_CONTEXT VM instruction.
Before, when the predicate contains one such clause, it set
``P_MFCONTEXT`` and inserted a I_CONTEXT for each subsequent
clause.  When using .qlf files though, the files may be loaded
in different order than when creating the .qlf files, leading
to clauses without an I_CONTEXT instruction to follow clauses
with this.