ADDED: "argnames" support.
This is experimental work to support "static dicts" as being discussed
in the PIP group.  This patch introduces the basic admin and extension
to read_term/2 and friends.

The naming is likely to change.  The functionality is close to ECLiPse
"struct" package.  For now we use the Ciao "argnames" as naming convention
because it is a rather unique name that can easily be changed into the
final name.   Functionality provided:

Read maps tag{k1:v1, ...} into a compound if _tag_ is declared to be
handled as "argnames".

Pedicates

  - :- argnames(:Term, +Options).
    E.g., argnames(point(x,y), []).   Options currently say whether the
    declaration is public to the module and what do do on redefinition.
  - arg_name(:NameOrTerm, ?Arg, ?Name)
    Translate between argument index (1..arity) and name
  - current_argnames(?Name, :Term)
    Query existing argname declarations.