ENHANCED: Index performance on common arguments.
Given a predicate with many clauses that has one or more arguments with
only a few possible values and (thus) many clauses in the hash buckets
for each of these values force fast failure when using a query with a
value that does not appear in that argument.

For example, suppose we have a business database that records the legal
status in one of the columns (arguments).   A query with a non-existing
lega status should fail quickly.   In the old system, the small number
of values could easily lead to a hash colission, causing the system to
try all clauses with the legal status of the collision.

This issue was reported by Abramo Bagnara