More robust check for `git` command
On MacOS the `git` command is only present after the (optional) xcode support has been installed. Prior to that there is an "imposter" `git` which outputs text and a dialog indicating how xcode can be acquired. 

For users with no need for xcode, this change provides a more robust check for an operational `git` by additionally checking for a man page. I think this change is platform agnostic but I haven't tested it on Unix or Windows.