FIXED: When a REPL loop receives the halt/0,1 exception, make it return.
Before, if non-threaded or the REPL loop runs in the main thread, terminate
the process.   Otherwise halt/0 was ignored.   The behaviour for the main
thread is still the same, but prolog/0 succeeds in other threads if the
thread calls halt/0.

Possibly we should add a flag to control this behaviour, i.e. whether or
not halt/0 inside prolog/0 terminates the process or merely the REPL loop.