ADDED: Implement read_term/1: implement reading attributed and labeled terms.
This patch supports Var{Attr: Value, ...}   if the Prolog flag `var_tag`
is set to `attvar`. This allows reading   terms with attributes. It also
supports _labeled subterms_ that allows to  read terms that are directed
graphs rather than trees.  This  includes   cyclic  terms.  For  this we
reserve the attributed `=`. For example

    ?- A = X{= : f(X)}.

Binds A to a cyclic term equivalent  to   executing  `X = f(X)`. This is
part of ongoing discussions in the PIP   working  group. It is currently
implemented by ECLiPSe and SWI-Prolog.